v1.0/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
boto3==1.34.158
botocore==1.34.158
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.8.0
distro==1.9.0
fastavro==1.9.5
filelock==3.15.4
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
protobuf==5.27.3
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.37.1
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
